Make SvcCache a simple Var.async[Addr.
Preiviously, the consul namer would eagerly start observing all services,
always. This works fine in small clusters but maybe be problematic in large
clusters.

We can fix this by creating a Var.async from the underlying service polling
loop.  Now, the namer will only query consul for services that linkerd is
trying to route on.
